moving heimdal queue config to common
