+++ kifi-backend/modules/common/conf/prod/heimdal.conf
@@ -14,8 +14,6 @@ statsd {
 
 dbplugin=disabled
 
-heimdal-events-queue-name = "heimdal-events"
-
 mongodb {
   heimdal {
     nodeA = "ds045508-a0.mongolab.com:45508"
