+++ kifi-backend/modules/common/conf/prod/prod.conf
@@ -1,17 +1,6 @@
 ## Config file for settings common to production servers
 
-java.source=1.6
-
-service {
- search {
-   host = "search"
-   port = 9000
- }
- shoebox {
-   host = "shoebox"
-   port = 9000
- }
-}
+java.source=1.7
 
 # Secret key
 # ~~~~~
@@ -88,12 +77,6 @@ cdn {
   bucket = "images-b-prod"
   base = "//djty7jcqog9qu.cloudfront.net"
 }
-
-#proxy = {
-#  host = localhost
-#  port = 8080
-#  protocol = http
-#}
 
 # result click tracker
 result-click-tracker = {
@@ -234,3 +217,8 @@ embedly {
 public-id {
   secret = "Or06kM3H447LGmNxWp5PGw==" // AES-128 key (base64 encoded)
 }
+
+# ~~~~~~~~~~~~~~~~~~~~~~
+# Shared SQS queues
+# ~~~~~~~~~~~~~~~~~~~~~~
+heimdal-events-queue-name = "heimdal-events"
