+++ kifi-backend/modules/heimdal/app/com/keepit/heimdal/HeimdalModule.scala
@@ -2,17 +2,14 @@ package com.keepit.heimdal
 
 import com.keepit.common.cache.CacheModule
 import com.keepit.inject.{CommonServiceModule, ConfigurationModule}
-import com.google.inject.{Provides, Singleton}
-import com.amazonaws.auth.BasicAWSCredentials
-import com.kifi.franz.QueueName
-import play.api.Play._
 import com.keepit.shoebox.ProdShoeboxServiceClientModule
 import com.keepit.social.RemoteSecureSocialModule
 
 abstract class HeimdalModule(
   // Common Functional Modules
   val cacheModule: CacheModule,
-  val mongoModule: MongoModule
+  val mongoModule: MongoModule,
+  val heimdalQueueModule: HeimdalQueueModule
 ) extends ConfigurationModule with CommonServiceModule  {
   // Service clients
   val shoeboxServiceClientModule = ProdShoeboxServiceClientModule()
