+++ kifi-backend/modules/heimdal/app/com/keepit/heimdal/HeimdalProdModule.scala
@@ -1,6 +1,5 @@
 package com.keepit.heimdal
 
-import com.keepit.common.cache.EhCacheCacheModule
 import com.keepit.inject.CommonProdModule
 import com.keepit.common.zookeeper.ProdDiscoveryModule
 import com.keepit.common.service.ServiceType
@@ -15,18 +14,11 @@ import com.amazonaws.regions.Regions
 
 case class HeimdalProdModule() extends HeimdalModule(
   cacheModule = HeimdalCacheModule(MemcachedCacheModule(), EhCacheCacheModule()),
-  mongoModule = ProdMongoModule()
+  mongoModule = ProdMongoModule(),
+  heimdalQueueModule = HeimdalQueueProdModule()
 ) with CommonProdModule {
   val discoveryModule = new ProdDiscoveryModule {
     def servicesToListenOn = ServiceType.SHOEBOX :: Nil
-  }
-
-  @Singleton
-  @Provides
-  def heimdalEventQueue(basicAWSCreds:BasicAWSCredentials): SQSQueue[HeimdalEvent] = {
-    val queueName = QueueName(current.configuration.getString("heimdal-events-queue-name").get)
-    val client = SimpleSQSClient(basicAWSCreds, Regions.US_WEST_1, buffered = false)
-    client.formatted[HeimdalEvent](queueName)
   }
 }
